# api/openapi.yaml
openapi: 3.1.0
info:
  title: CoachPulse API
  version: "1.0.0"
  description: |
    The API behind CoachPulse — a business operating system for independent
    personal trainers.

    Two conventions run through everything here and are worth reading before
    the endpoints.

    **Money is integer minor units.** Every amount is a whole number of cents,
    fils or agorot, paired with an ISO-4217 code. There are no floats. A
    ledger that rounds is a ledger that cannot be reconciled.

    **Errors carry a stable machine code.** Clients branch on `error.code`,
    never on `error.message`: the message is for humans and may be reworded,
    the code is part of the contract. `insufficient_credits` carries the
    remaining balance in `meta`, which is what lets the app offer a renewal
    at the moment a pack runs out.

    Measurements follow money's lead: loads are grams, RPE is tenths (85 is
    RPE 8.5), body fat is basis points. Display units are the client's choice;
    the stored value never rounds.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.coachpulse.io
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /v1/settings:
    get:
      tags: [Settings]
      summary: The practice's settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
    patch:
      tags: [Settings]
      summary: Change settings
      description: |
        Owner only. Fields left out are left alone.

        Setting `country` also sets `week_start` to that country's (Sunday in
        Saudi Arabia, Monday in the UAE) unless the same change names one.
        `currency` answers `409 currency_locked` once money has been recorded.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsPatch" }
      responses:
        "200":
          description: The settings as they now stand
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /v1/clients:
    get:
      tags: [Clients]
      summary: List clients
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ClientStatus" } }
        - { name: search, in: query, schema: { type: string }, description: Matches name, email and phone }
        - { name: tag_id, in: query, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: object
                properties:
                  clients:
                    type: array
                    items: { $ref: "#/components/schemas/Client" }
    post:
      tags: [Clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /v1/sessions:
    get:
      tags: [Scheduling]
      summary: The calendar over a window
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session_" }
    post:
      tags: [Scheduling]
      summary: Book a session
      description: |
        Refused with `scheduling_conflict` if the slot overlaps an existing
        session including the tenant's buffer. The check is a database
        exclusion constraint, so two devices syncing the same booking cannot
        both win.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_type_id, starts_at]
              properties:
                session_type_id: { type: string, format: uuid }
                starts_at: { type: string, format: date-time }
                client_ids:
                  type: array
                  items: { type: string, format: uuid }
                location: { type: string }
                notes: { type: string }
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session_" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /v1/invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      parameters:
        - { name: client_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/InvoiceStatus" } }
        - { name: outstanding, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items: { $ref: "#/components/schemas/Invoice" }
    post:
      tags: [Invoices]
      summary: Create a draft
      description: A draft posts nothing and takes no number, so deleting one
        leaves no gap in the numbering.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceDraftInput" }
      responses:
        "201":
          description: Drafted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
...
  /v1/payment-methods:
    get:
      tags: [Invoices]
      summary: How this trainer accepts settlement
      responses:
        "200":
          description: Methods
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment_methods:
                    type: array
                    items: { $ref: "#/components/schemas/PaymentMethod" }
    post:
      tags: [Invoices]
      summary: Add a settlement method
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentMethod" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentMethod" }
  /v1/exercises:
    get:
      tags: [Programming]
      summary: The exercise library
      description: Seeded per tenant at signup, organised by movement pattern.
      parameters:
        - { name: category, in: query, schema: { $ref: "#/components/schemas/ExerciseCategory" } }
        - { name: search, in: query, schema: { type: string }, description: Matches name, muscle and equipment }
      responses:
        "200":
          description: Exercises
          content:
            application/json:
              schema:
                type: object
                properties:
                  exercises:
                    type: array
                    items: { $ref: "#/components/schemas/Exercise" }
    post:
      tags: [Programming]
      summary: Add a custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category]
              properties:
                name: { type: string }
                category: { $ref: "#/components/schemas/ExerciseCategory" }
                primary_muscle: { type: string }
                equipment: { type: string }
                instructions: { type: string }
                video_url: { type: string }
                demo_media_id: { type: string, format: uuid, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Exercise" }
...
  /v1/programs:
    get:
      tags: [Programming]
      summary: List programmes
      parameters:
        - { name: templates, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Programmes
          content:
            application/json:
              schema:
                type: object
                properties:
                  programs:
                    type: array
                    items: { $ref: "#/components/schemas/Program" }
    post:
      tags: [Programming]
      summary: Create a programme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                is_template: { type: boolean }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
...
  /v1/workouts:
    post:
      tags: [Programming]
      summary: Open a workout to log sets against
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id]
              properties:
                client_id: { type: string, format: uuid }
                day_id: { type: string, format: uuid, nullable: true }
                session_id: { type: string, format: uuid, nullable: true }
                week_number: { type: integer, default: 1 }
                performed_on: { type: string, format: date }
      responses:
        "201":
          description: Opened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workout" }
...
  /v1/waivers:
    get:
      tags: [Clients]
      summary: Waiver documents
      parameters:
        - { name: active, in: query, schema: { type: boolean, default: true } }
      responses:
        "200":
          description: Waivers
          content:
            application/json:
              schema:
                type: object
                properties:
                  waivers:
                    type: array
                    items: { $ref: "#/components/schemas/Waiver" }
    post:
      tags: [Clients]
      summary: Publish a waiver
      description: A waiver with the same title becomes a new version; the old
        one stays attached to the signatures made against it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title: { type: string }
                body: { type: string }
      responses:
        "201":
          description: Published
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Waiver" }

  # -------------------------------------------------------------------------
  # Media
  # -------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, minLength: 8, maxLength: 255 }
      description: |
        Makes a mutating request safe to retry, which is what lets the offline
        outbox retry aggressively. A replay returns the stored response
        verbatim with `Idempotent-Replay: true`. Reusing a key with a
        *different* body is refused with `idempotency_key_reused`.
  responses:
    Forbidden:
      description: |
        Not allowed. `feature_not_in_plan` (meta: `feature`, `plan`) and
        `plan_limit_reached` (meta: `limit`, `max`, `plan`) are the ones the
        app answers with an upgrade prompt.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    BadRequest:
      description: Malformed or failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Conflict:
      description: A state precondition was violated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Stable and part of the contract — branch on this, not on the
                message.
              examples: [insufficient_credits, overpayment, scheduling_conflict,
                         invoice_not_payable, invalid_state_transition,
                         idempotency_key_reused, validation_failed]
            message: { type: string, description: For humans; may be reworded }
            fields:
              type: object
              additionalProperties: { type: string }
              description: Per-field validation detail
            meta:
              type: object
              additionalProperties: true
              description: |
                Structured context the client can act on. An
                `insufficient_credits` refusal carries `remaining` and
                `required`, which is what drives the renewal prompt.
            request_id: { type: string }
...
    WorkingHours:
      type: object
      description: Weekday ("0" = Sunday … "6") to "HH:MM" intervals. A day
        absent is a day off.
      additionalProperties:
        type: array
        items:
          type: array
          items: { type: string, examples: ["06:00"] }
          minItems: 2
          maxItems: 2
    Targets:
      type: object
      properties:
        monthly_revenue_minor: { type: integer, format: int64 }
        weekly_sessions: { type: integer }
        active_clients: { type: integer }
    Automations:
      type: object
      properties:
        renewal_due: { type: boolean }
        overdue_invoice: { type: boolean }
        inactive_client: { type: boolean }
        programme_ending: { type: boolean }
        inactive_after_days: { type: integer }
    Settings:
      type: object
      properties:
        business_name: { type: string }
        currency: { type: string }
        currency_locked: { type: boolean, description: True once money has been recorded }
        timezone: { type: string, examples: [Asia/Dubai] }
        country: { type: [string, "null"], examples: [AE] }
        language: { type: string, enum: [en, ar] }
        document_language: { type: string, enum: [en, ar, bilingual] }
        digits: { type: string, enum: [latn, arab] }
        week_start: { type: integer, minimum: 0, maximum: 6 }
        working_hours: { $ref: "#/components/schemas/WorkingHours" }
        targets: { $ref: "#/components/schemas/Targets" }
        automations: { $ref: "#/components/schemas/Automations" }
        session_timeout_days: { type: integer }
        buffer_minutes: { type: integer }
        allow_overdraft: { type: boolean }
        no_show_is_billable: { type: boolean }
        low_balance_threshold: { type: integer }
    SettingsPatch:
      type: object
      properties:
        business_name: { type: string }
        currency: { type: string }
        timezone: { type: string }
        country: { type: string }
        language: { type: string, enum: [en, ar] }
        document_language: { type: string, enum: [en, ar, bilingual] }
        digits: { type: string, enum: [latn, arab] }
        week_start: { type: integer, minimum: 0, maximum: 6 }
        working_hours: { $ref: "#/components/schemas/WorkingHours" }
        targets: { $ref: "#/components/schemas/Targets" }
        automations: { $ref: "#/components/schemas/Automations" }
        session_timeout_days: { type: integer, minimum: 1, maximum: 365 }
        buffer_minutes: { type: integer, minimum: 0, maximum: 240 }
        allow_overdraft: { type: boolean }
        no_show_is_billable: { type: boolean }
        low_balance_threshold: { type: integer, minimum: 0, maximum: 100 }
...
    ClientStatus:
      type: string
      enum: [lead, active, paused, archived]
    Client:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        date_of_birth: { type: string, format: date, nullable: true }
        status: { $ref: "#/components/schemas/ClientStatus" }
        emergency_contact_name: { type: string, nullable: true }
        emergency_contact_phone: { type: string, nullable: true }
        medical_notes:
          type: string
          nullable: true
          description: Only present when requested with `include_medical=true`.
        allow_overdraft: { type: boolean, nullable: true }
        default_rate_minor: { type: integer, nullable: true }
        notes: { type: string }
        tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
        server_seq: { type: integer, format: int64 }
    ClientInput:
      type: object
      properties:
        full_name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        date_of_birth: { type: string, format: date, nullable: true }
        status: { $ref: "#/components/schemas/ClientStatus" }
        emergency_contact_name: { type: string, nullable: true }
        emergency_contact_phone: { type: string, nullable: true }
        medical_notes: { type: string, nullable: true }
        allow_overdraft: { type: boolean, nullable: true }
        default_rate_minor: { type: integer, nullable: true }
        notes: { type: string }
        tag_ids:
          type: array
          items: { type: string, format: uuid }
    Tag:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        colour: { type: string, nullable: true }
...
    Session_:
      type: object
      description: A booked slot on the calendar.
      properties:
        id: { type: string, format: uuid }
        session_type_id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        status: { type: string, enum: [scheduled, cancelled] }
        location: { type: string }
        notes: { type: string }
        attendees:
          type: array
          items: { $ref: "#/components/schemas/Attendee" }
        server_seq: { type: integer, format: int64 }
    AttendanceStatus:
      type: string
      enum: [scheduled, completed, late_cancel, early_cancel, no_show]
    Attendee:
      type: object
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        client_name: { type: string }
        status: { $ref: "#/components/schemas/AttendanceStatus" }
        credits_charged: { type: integer }
        marked_at: { type: string, format: date-time, nullable: true }
        notes: { type: string }
...
    InvoiceStatus:
      type: string
      enum: [draft, issued, partially_paid, settled, void]
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        client_name: { type: string }
        number: { type: string, nullable: true, description: Assigned at issue; gap-free }
        status: { $ref: "#/components/schemas/InvoiceStatus" }
        currency: { type: string }
        total_minor: { type: integer, format: int64 }
        paid_minor: { type: integer, format: int64 }
        balance_minor: { type: integer, format: int64 }
        issue_date: { type: string, format: date, nullable: true }
        due_date: { type: string, format: date, nullable: true }
        is_overdue:
          type: boolean
          description: Derived from the due date and balance, never stored.
        days_overdue: { type: integer }
        notes: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/InvoiceLine" }
        payment_instructions:
          $ref: "#/components/schemas/PaymentInstructions"
          description: |
            Frozen at issue. An invoice keeps the IBAN it was sent with even if
            the trainer later changes banks.
        journal_entry_id: { type: string, format: uuid, nullable: true }
        issued_at: { type: string, format: date-time, nullable: true }
        settled_at: { type: string, format: date-time, nullable: true }
        voided_at: { type: string, format: date-time, nullable: true }
        void_reason: { type: string }
        has_share_link: { type: boolean }
        server_seq: { type: integer, format: int64 }
    PaymentInstructions:
      type: object
      nullable: true
      properties:
        methods:
          type: array
          items: { $ref: "#/components/schemas/PaymentMethod" }
        reference:
          type: string
          description: What the client should write on the transfer — the invoice number.
        note: { type: string }
    InvoiceLine:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          enum: [package, service]
          description: |
            A package line sells prepaid credits and credits Deferred Revenue;
            a service line bills work already done and credits revenue
            directly.
        description: { type: string }
        quantity: { type: integer }
        unit_price_minor: { type: integer, format: int64 }
        amount_minor: { type: integer, format: int64 }
        package_credits: { type: integer, nullable: true }
        credits_expire_on: { type: string, format: date, nullable: true }
        sort_order: { type: integer }
    InvoiceDraftInput:
      type: object
      required: [client_id, lines]
      properties:
        client_id: { type: string, format: uuid }
        currency: { type: string }
        due_date: { type: string, format: date, nullable: true }
        notes: { type: string }
        lines:
          type: array
          minItems: 1
          items:
            type: object
            required: [description, quantity, unit_price_minor]
            properties:
              kind: { type: string, enum: [package, service], default: service }
              description: { type: string }
              quantity: { type: integer, minimum: 1 }
              unit_price_minor: { type: integer, minimum: 0 }
              package_credits:
                type: integer
                nullable: true
                description: >-
                  Credits granted per unit; the pack is worth
                  quantity x package_credits. Ten sessions at 50.00 each is
                  quantity 10, unit_price_minor 5000, package_credits 1 — not
                  package_credits 10, which grants a hundred.
                  Required on a package line, forbidden on a service line.
              credits_expire_on: { type: string, format: date, nullable: true }
...
    PaymentMethod:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          enum: [bank_transfer, digital_wallet, cash, cheque, other]
        label: { type: string }
        details:
          type: object
          description: |
            Structured settlement fields. Which ones matter varies by country
            and by app.
          properties:
            account_holder: { type: string }
            bank_name: { type: string }
            iban: { type: string }
            swift_bic: { type: string }
            account_number: { type: string }
            routing_number: { type: string }
            handle: { type: string, description: A P2P tag or link }
        instructions: { type: string }
        is_default: { type: boolean }
...
    ExerciseCategory:
      type: string
      description: Movement pattern, not muscle — how a coach checks a session
        is balanced.
      enum: [squat, hinge, lunge, horizontal_push, vertical_push,
             horizontal_pull, vertical_pull, carry, core, olympic,
             conditioning, mobility, other]
    Exercise:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        category: { $ref: "#/components/schemas/ExerciseCategory" }
        primary_muscle: { type: string }
        equipment: { type: string }
        instructions: { type: string }
        video_url:
          type: string
          description: YouTube or Vimeo only — allowlisted, not format-checked.
        is_custom: { type: boolean }
    Program:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        is_template: { type: boolean }
        total_weeks: { type: integer }
        blocks:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              weeks:
                type: integer
                description: |
                  The block's days repeat for this many weeks rather than being
                  stored once per week, so editing week one edits the block.
              days:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    name: { type: string }
                    prescriptions:
                      type: array
                      items: { $ref: "#/components/schemas/Prescription" }
...
    Workout:
      type: object
      properties:
        id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        day_id: { type: string, format: uuid, nullable: true }
        day_name: { type: string }
        session_id: { type: string, format: uuid, nullable: true }
        week_number: { type: integer }
        performed_on: { type: string, format: date }
        status: { type: string, enum: [in_progress, completed, skipped] }
        notes: { type: string }
        sets:
          type: array
          items: { $ref: "#/components/schemas/SetLog" }
        server_seq: { type: integer, format: int64 }
...
    Waiver:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        version: { type: integer }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
...